requirements_minimal.txt
streamlit
tensorflow-cpu
numpy
Pillow
scikit-image
requests